[jin.100] 인증 관리 Step1, 2 (#9)
* docs: 1단계 기능 명세 추가

* feat : 토큰 발급하는 API 생성

* feat : 토큰 발급하는 API 생성

* docs : 2단계 요구사항 추가

* feat : AuthenticationPrincipalArgumentResolver 추가

* feat : accessToken이 있는 유저만 예약 추가/삭제 가능하도록 수정

* feat : accessToken이 있는 유저만 예약 추가/삭제 가능하도록 수정

* feat : 예약 삭제 시 본인의 예약만 삭제 가능하도록 수정

* feat : reservation 테이블에 member_id 필드 추가

* refactor : ArgumentResolver에서 username 대신 loginMember 반환하도록 변경

* refactor : AuthenticationPrincipal 어노테이션 이름 LoginUser로 변경

* fix : member_id 필드 추가 후 Reservation를 찾아올 수 없던 sql 에서 수정

* test: 다른 사람의 예약 삭제 불가능한 테스트케이스 토큰 기반으로 판별하도록 변경

* refactor: Member에 default constructor 추가

* test : 멤버 테스트하는 클래스 이름 MemberE2ETest 로 변경

* fix : `@LoginUser` 어노테이션이 있지만 Authorization 헤더가 없을 떄 UnAuthorizedException 던지도록 수정

* test : 개인정보 조회 기능 테스트케이스 추가

* test : 토큰이 없이 인증 필요한 API에 요청보내는 테스트케이스 추가

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,13 +1,18 @@
 package nextstep.reservation;
 
+import nextstep.auth.LoginMember;
+import nextstep.member.Member;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.ForbiddenAccessException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ReservationService {
@@ -15,15 +20,23 @@ public class ReservationService {
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public final MemberDao memberDao;
+
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, LoginMember loginMember) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
+            throw new NullPointerException();
+        }
+
+        Member member = memberDao.findById(loginMember.getId());
+        if (member == null) {
             throw new NullPointerException();
         }
 
@@ -34,6 +47,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
+                member,
                 reservationRequest.getName()
         );
 
@@ -49,10 +63,14 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, LoginMember loginMember) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
+        }
+
+        if(!Objects.equals(reservation.getMember().getUsername(), loginMember.getUsername())){
+            throw new ForbiddenAccessException();
         }
 
         reservationDao.deleteById(id);
